% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-profiles.R
\name{create_profiles}
\alias{create_profiles}
\title{Create attribute profiles}
\usage{
create_profiles(attributes)
}
\arguments{
\item{attributes}{The number of attributes.}
}
\description{
Create attribute profiles
}
