% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{dcmApp}
\alias{dcmApp}
\title{Run the LCDM Item Response Function App}
\usage{
dcmApp(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\value{
A shiny app.
}
\description{
Run the LCDM Item Response Function App
}
