% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm-probs-package.R
\docType{package}
\name{dcmprobs-package}
\alias{dcmprobs}
\alias{dcmprobs-package}
\title{dcmprobs: Log-linear Cognitive Diagnostic Model Item Response Function}
\description{
A shiny app for illustrating item response probabilities for each attribute profile. Users input parameter values for relevant variables. The Shiny app then calculate the probability that an individual in each class provides a correct response to an item with the given parameter values.
}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

}
\keyword{internal}
